change visibility of fixed_bytes_str module to private
move its test to its own module file.
