+++ src/lib.rs
@@ -1,2 +1,2 @@
-pub mod fixed_bytes_str;
+mod fixed_bytes_str;
 pub mod tokenizer;
